requirements.txt
numpy
torch
torchaudio
pytorch_lightning
hydra-core
wandb
streamlit
dvc
python-dotenv
black
matplotlib
plotly
librosa
pyspark